+++ changelog.md
@@ -16,8 +16,9 @@ added:
   * config, colors and styles were added to this in commands
   * terminalApi has now styles and colors getters
   * verApi has now a branch, edition and isExperimental
-  * filter in command Collection
+  * filter and exclude are now in command Collection
   * cleanReturner -> allows to simply return the array if needed
+  * new pipe: '<<'    -> from a file append to current result
 fixed:
   * multiline commands are now properly supported as new typing system was introduced
   * wrongly written inputbox in logs
@@ -26,6 +27,7 @@ fixed:
 added commands:
   * base64 -> decodes and encodes base64 data
   * color -> colorizes data
+  * call -> calls a function
 
 There's now a suggestion about commands that were not found